Fix: Fixed Code Launcher docker handler and other issues.
Fixed the logic for the docker handler - now it's no longer responsability of the launcher to build the image before the code execution. The launcher will use the already built image in the system and return a error if there's no image.

The launcher will no longer copy the data through the image building. Now it will copy the files directly to the container using the "docker cp" feature.

Changed System.out messages to log implementation.