Fix: Fixed small errors.
Errors output was not returning any data because it was checking if the file inside docker would exist, instead of checking if the local file exists.

The id in the report should be set by the CodeAnalysisOutputEntity because the analysis output will be always present. The CodeExecutionResultEntity may not be present if a syntax error was found, for example.